clearCache {R.cache}	R Documentation

Removes all files in a cache file directory

Description

Removes all files in a cache file directory.

Usage

## Default S3 method:
clearCache(path=getCachePath(...), ..., recursive=FALSE, prompt=TRUE && interactive())

Arguments

path
A character string specifying the directory to be cleared. By default, the path is what is returned by getCachePath() which arguments ... are also passed.

...
Arguments passed to getCachePath(), especially argument dirs to specify subdirectories.

recursive
If TRUE, subdirectories are also removed, otherwise just the files in the specified directory.

prompt
If TRUE, the user will be prompted to confirm that the directory will cleared before files are removed.

Details

If the specified directory does not exists, an exception is thrown.

Value

Returns (invisibly) a character vector of pathnames of the files removed. If no files were removed, NULL is returned.

Author(s)

Henrik Bengtsson

[Package R.cache version 0.17.0 Index]