findCache {R.cache}	R Documentation

Locates a cache file

Description

Locates a cache file from a key object.

Usage

## Default S3 method:
findCache(key=NULL, ...)

Arguments

key
An optional object from which a hexadecimal hash code will be generated and appended to the filename.

...
Additional argument passed to generateCache().

Value

Returns the pathname as a character, or NULL if the no cached data exists.

Author(s)

Henrik Bengtsson

See Also

generateCache(). loadCache().

[Package R.cache version 0.17.0 Index]