generateCache {R.cache}	R Documentation

Generates a cache pathname from a key object

Description

Generates a cache pathname from a key object.

Usage

## Default S3 method:
generateCache(key, suffix=".Rcache", ...)

Arguments

key
A list or an environment from which a character string checksum will be calculated and that will constitute the name part of the cache filename.

suffix
A character string to be appended to the end of the filename.

...
Arguments passed to getCachePath().

Value

Returns the pathname as a character string.

Author(s)

Henrik Bengtsson

See Also

findCache(). Internally, the generic function getChecksum() is used to calculate the checksum of argument key.

[Package R.cache version 0.17.0 Index]