getCachePath {R.cache}	R Documentation

Gets the path to the file cache directory

Description

Gets the path to the file cache directory. If missing, the directory is created.

Usage

## Default S3 method:
getCachePath(dirs=NULL, path=NULL, rootPath=getCacheRootPath(), ...)

Arguments

dirs
A character vector constituting the path to the cache subdirectory (of the cache root directory as returned by getCacheRootPath()) to be used. If NULL, the path will be the cache root path.

path, rootPath
(Advanced) character strings specifying the explicit/default cache path and root cache path.

...
Not used.

Value

Returns the path as a character string. If the user does not have write permissions to the path, then an error is thrown.

Author(s)

Henrik Bengtsson

See Also

setCachePath.

[Package R.cache version 0.17.0 Index]