getCacheRootPath {R.cache}	R Documentation

Gets the root path to the file cache directory

Description

Gets the root path to the file cache directory.

Usage

## Default S3 method:
getCacheRootPath(defaultPath=NULL, ...)

Arguments

defaultPath
The default path, if no user-specified directory has been given.

...
Not used.

Value

Returns the path as a character string.

Author(s)

Henrik Bengtsson

See Also

Too set the directory where cache files are stored, see setCacheRootPath().

Examples

  print(getCacheRootPath())

[Package R.cache version 0.17.0 Index]