setupCacheRootPath {R.cache}	R Documentation

Interactively offers the user to set up the default root path

Description

Interactively offers the user to set up the default root path.

Usage

## Default S3 method:
setupCacheRootPath(defaultPath=NULL, ...)

Arguments

defaultPath
Default root path to set.

...
Not used.

Details

If the cache root path is already set, it is used and nothing is done. If the "default" root path (defaultPath) exists, it is used, otherwise, if running interactively, the user is asked to approve the usage (and creation) of the default root path. In all other cases, the cache root path is set to a session-specific temporary directory.

Value

Returns (invisibly) the root path, or NULL if running a non-interactive session.

Author(s)

Henrik Bengtsson

See Also

Internally, setCacheRootPath() is used to set the cache root path. The interactive() function is used to test whether R is running interactively or not.

[Package R.cache version 0.17.0 Index]