Options used by R.cache

Below are all R options specific to the R.cache package.
WARNING: Note that the names and the default values of these options may change in future versions of the package. Please use with care until further notice.

Options for controlling futures

R.cache.compress:

If TRUE, saveCache() will write compressed cache files, otherwise not. (Default: FALSE)

R.cache.enabled:

If TRUE, loadCache() is reading from and saveCache() is writing to the cache, otherwise not. (Default: TRUE)

R.cache.rootPath:

A character string specifying the default cache root path. If not set, environment variable R_CACHE_ROOTPATH is considered.

R.cache.touchOnLoad:

If TRUE, loadCache() will update the "last-modified" timestamp of the cache file (to the current time), otherwise not. (Default: FALSE)